+++ api/src/main/java/com/wancs/battle_run/domain/running/dto/request/UpdateRecordRequestDto.java
@@ -15,7 +15,7 @@ public class UpdateRecordRequestDto {
 
     private Long runningTime;
 
-    private Double pace;
+    private Integer pace;
 
     private Integer calorie;
